refactor(gatt): Move ClientAttTable into its own module
